Address initial PR (#38) feedback.
