+++ src/nav2_navigation/bt_navigator/src/main.cpp
@@ -15,7 +15,7 @@
 #include <string>
 #include <memory>
 #include "rclcpp/rclcpp.hpp"
-#include "bt_navigator/BtNavigator.hpp"
+#include "bt_navigator/bt_navigator.hpp"
 
 int main(int argc, char ** argv)
 {
