+++ src/nav2_tasks/include/nav2_tasks/ComputePathToPoseTaskServer.hpp
@@ -15,7 +15,7 @@
 #ifndef NAV2_TASKS__COMPUTEPATHTOPOSETASKSERVER_HPP_
 #define NAV2_TASKS__COMPUTEPATHTOPOSETASKSERVER_HPP_
 
-#include "nav2_tasks/TaskServer.hpp"
+#include "nav2_tasks/task_server.hpp"
 #include "nav2_tasks/ComputePathToPoseTaskMessages.hpp"
 
 namespace nav2_tasks
