+++ src/nav2_tasks/include/nav2_tasks/FollowPathTaskServer.hpp
@@ -15,7 +15,7 @@
 #ifndef NAV2_TASKS__FOLLOWPATHTASKSERVER_HPP_
 #define NAV2_TASKS__FOLLOWPATHTASKSERVER_HPP_
 
-#include "nav2_tasks/TaskServer.hpp"
+#include "nav2_tasks/task_server.hpp"
 #include "nav2_tasks/FollowPathTaskMessages.hpp"
 
 namespace nav2_tasks
