+++ src/nav2_tasks/include/nav2_tasks/task_status.hpp
@@ -12,25 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DWA_CONTROLLER__DWACONTROLLER_HPP_
-#define DWA_CONTROLLER__DWACONTROLLER_HPP_
+#ifndef NAV2_TASKS__TASK_STATUS_HPP_
+#define NAV2_TASKS__TASK_STATUS_HPP_
 
-#include <string>
-#include "nav2_tasks/FollowPathTaskServer.hpp"
-
-namespace dwa_controller
+namespace nav2_tasks
 {
 
-class DwaController : public nav2_tasks::FollowPathTaskServer
-{
-public:
-  explicit DwaController(const std::string & name);
-  DwaController() = delete;
-  ~DwaController();
+typedef enum { SUCCEEDED, FAILED, RUNNING, CANCELED } TaskStatus;
 
-  nav2_tasks::TaskStatus executeAsync(const nav2_tasks::FollowPathCommand::SharedPtr path) override;
-};
+}  // namespace nav2_tasks
 
-}  // namespace dwa_controller
-
-#endif  // DWA_CONTROLLER__DWACONTROLLER_HPP_
+#endif  // NAV2_TASKS__TASK_STATUS_HPP_
