+++ src/nav2_util/package.xml
@@ -4,8 +4,7 @@
   <name>nav2_util</name>
   <version>0.1.0</version>
   <description>TODO</description>
-  <author email="oregon.robotics.team@intel.com">Oregon Robotics Team</author>
-  <maintainer email="oregon.robotics.team@intel.com">Oregon Robotics Team</maintainer>
+  <maintainer email="michael.jeronimo@intel.com">Michael Jeronimo</maintainer>
   <license>Apache License 2.0</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
