+++ src/nav2_world_model/src/main.cpp
@@ -14,7 +14,7 @@
 
 #include <memory>
 #include "rclcpp/rclcpp.hpp"
-#include "nav2_world_model/WorldModel.hpp"
+#include "nav2_world_model/world_model.hpp"
 
 int main(int argc, char ** argv)
 {
